requirements.txt
DAWG-Python==0.7.2
docopt==0.6.2
mysql-connector-python==9.3.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
